--- components/GetUserCanEdit.tsx
@@ -1,11 +1,0 @@
-import { vsmProject } from "../interfaces/VsmProject";
-
-export function getUserCanEdit(
-  account: {
-    username: string;
-  },
-  project: vsmProject
-): boolean {
-  const userName = account?.username.split("@")[0];
-  return project?.created.userIdentity === userName;
-}
