+++ utils/StripGarbage.ts
