+++ utils/getMyAccess.tsx
@@ -15,10 +15,13 @@ export function getMyAccess(
 
   //If we are the creator of the project, return Admin
   const shortName = getUserShortName(account);
-  if (project?.created?.userIdentity === shortName) return "Admin";
+  if (project?.created?.userIdentity.toUpperCase() === shortName)
+    return "Admin";
 
   //Else, check if we have been given a role, then return that role
-  const found = project?.userAccesses.find((u) => u.user === shortName);
+  const found = project?.userAccesses.find(
+    (u) => u.user.toUpperCase() === shortName
+  );
   if (found) return found.role;
 
   //If not given a role, we default to "Reader"
