+++ utils/getUserShortName.tsx
@@ -1,2 +1,2 @@
 export const getUserShortName = (account: { username: string }): string =>
-  account.username.split("@")[0];
+  account.username.split("@")[0].toUpperCase();
